+++ flutter_inappwebview_platform_interface/CHANGELOG.md
@@ -3,6 +3,7 @@
 - Added `containerId` and `proxySettings` properties to `InAppWebViewSettings`
 - Added `PlatformContainerController` class
 - Added `WebViewFeature.MULTI_PROFILE`
+- Added Linux support to `PlatformContainerController`, `InAppWebViewSettings.containerId` and `InAppWebViewSettings.proxySettings`
 - Updated `flutter_inappwebview_internal_annotations` dependency from `^1.2.0` to `^1.3.0`
 - Added `isClassSupported`, `isPropertySupported`, `isMethodSupported` static methods for all main classes, such as `PlatformInAppWebViewController`, `InAppWebViewSettings`, `PlatformInAppBrowser`, etc., in order to check if a class, property, or method is supported by the platform at runtime
 - Added `isSupported` method to all custom enum classes